@@ -3,29 +3,30 @@
    "version": "0.1.0",
    "private": true,
    "dependencies": {
-      "@reduxjs/toolkit": "^1.9.0",
-      "@testing-library/jest-dom": "^5.12.0",
-      "@testing-library/react": "^11.2.6",
-      "@testing-library/user-event": "^12.8.3",
-      "axios": "^0.21.1",
-      "framer-motion": "^4.1.11",
-      "react": "^17.0.2",
+      "@reduxjs/toolkit": "^1.9.1",
+      "@testing-library/jest-dom": "^5.16.5",
+      "@testing-library/react": "^13.4.0",
+      "@testing-library/user-event": "^14.4.3",
+      "axios": "^1.2.1",
+      "framer-motion": "^7.10.2",
+      "react": "^18.2.0",
       "react-content-loader": "^6.2.0",
-      "react-dom": "^17.0.2",
-      "react-lazy-load-image-component": "^1.5.1",
-      "react-player": "^2.9.0",
+      "react-dom": "^18.2.0",
+      "react-lazy-load-image-component": "^1.5.6",
+      "react-player": "^2.11.0",
       "react-redux": "^8.0.5",
-      "react-router-dom": "^5.2.0",
-      "react-scripts": "4.0.3",
-      "react-slick": "^0.28.1",
-      "screenfull": "^5.1.0",
+      "react-router-dom": "^6.5.0",
+      "react-scripts": "^5.0.1",
+      "react-slick": "^0.29.0",
+      "react-toastify": "^9.1.1",
+      "screenfull": "^6.0.2",
       "slick-carousel": "^1.8.1",
-      "styled-components": "^5.2.3",
-      "web-vitals": "^1.1.1"
+      "styled-components": "^5.3.6",
+      "web-vitals": "^3.1.0"
    },
    "scripts": {
-      "start": "react-scripts --openssl-legacy-provider start",
-      "build": "react-scripts --openssl-legacy-provider build",
+      "start": "react-scripts start",
+      "build": "react-scripts build",
       "test": "react-scripts test",
       "eject": "react-scripts eject"
    },
