@@ -1,15 +1,19 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
+import 'react-toastify/dist/ReactToastify.css';
 import { Provider } from 'react-redux';
 import { store } from './store/store';
+import { BrowserRouter as Router } from 'react-router-dom';
 
-ReactDOM.render(
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
    <React.StrictMode>
-      <Provider store={store}>
-         <App />
-      </Provider>
+      <Router>
+         <Provider store={store}>
+            <App />
+         </Provider>
+      </Router>
    </React.StrictMode>,
-   document.getElementById('root'),
 );
