@@ -2,14 +2,12 @@ const mongoose = require('mongoose');
 
 let url;
 if (process.env.NODE_ENV) {
-	url = process.env.ATLAS;
+   url = process.env.ATLAS;
 } else {
-	url = require('./atlas');
+   url = require('./atlas');
 }
-
+mongoose.set('strictQuery', false);
 mongoose.connect(url, {
-	useNewUrlParser: true,
-	useCreateIndex: true,
-	useFindAndModify: false,
-	useUnifiedTopology: true,
+   useNewUrlParser: true,
+   useUnifiedTopology: true,
 });
